requirements.txt
fastapi
uvicorn[standard]
python-multipart
python-dotenv
pydantic

pinecone
sentence-transformers
numpy
langchain-core
langchain-google-genai

PyPDF2
Pillow
pdf2image
python-docx

google-cloud-vision
google-cloud-storage
langgraph
